moved files related to customize into filter folder
